refactor(notebook-cloud): split viewer route views, leaf components, hooks out of index.tsx
viewer/index.tsx was a 3,136-line entry that mixed routing, three route
views, the notebook viewer, auth hooks, presence, and a pile of helpers.
Pull the loosely-coupled layer into focused modules matching the viewer/
convention:

- cloud-viewer-types: shared view types
- cloud-viewer-config: config/bootstrap loaders + route path helpers
- use-cloud-auth: useCloudPrototypeAuth / AppSessionStatus / AppSessionBridge
- home-view / oidc-callback-view / notebook-list-view: the three route views
- cloud-notebook-dashboard-view: the dashboard presentation family
- cloud-presence-status / cloud-notebook-title / cloud-auth-controls: leaf chrome

index.tsx keeps App, CloudNotebookProviders, NotebookViewer, and mount; drops
3,136 -> 1,343 lines. Pure code motion, no behavior change. Typecheck clean;
the shared-shell source guardrails (now corpus-based) stay green.
