+++ apps/notebook-cloud/viewer/notebook-list-view.tsx
@@ -1,0 +1,511 @@
+import { useCallback, useEffect, useMemo, useState, type FormEvent } from "react";
+import {
+  AlertCircle,
+  BookOpen,
+  FilePlus2,
+  KeyRound,
+  Loader2,
+  LogOut,
+  RotateCcw,
+} from "lucide-react";
+import { useTheme } from "@/hooks/useTheme";
+import { clearCloudAppSession } from "./app-session";
+import { CloudNotebookSignInButton } from "./cloud-auth-controls";
+import { CloudNotebookDashboard } from "./cloud-notebook-dashboard-view";
+import { loadCloudNotebookListBootstrap } from "./cloud-viewer-config";
+import {
+  clearCloudPrototypeDevAuth,
+  fetchWithCloudPrototypeAuth,
+  type CloudPrototypeAuthState,
+} from "./collaborator-auth";
+import { cloudResponseError } from "./cloud-response";
+import { projectCloudNotebookDashboard, type CloudNotebookListItem } from "./notebook-dashboard";
+import {
+  clearCachedCloudNotebookList,
+  readCachedCloudNotebookList,
+  writeCachedCloudNotebookList,
+} from "./notebook-list-cache";
+import { applyDocumentTheme, CLOUD_VIEWER_THEME_STORAGE_KEY } from "./theme";
+import {
+  useCloudAppSessionBridge,
+  useCloudAppSessionStatus,
+  useCloudPrototypeAuth,
+} from "./use-cloud-auth";
+import type {
+  CloudNotebookCreateResponse,
+  CloudNotebookListBootstrap,
+  CloudNotebookListResponse,
+  CloudNotebookListState,
+  CloudNotebookRenameState,
+  CloudNotebookUpdateResponse,
+  CloudViewerAuthConfig,
+} from "./cloud-viewer-types";
+
+export function CloudNotebookListView({ authConfig }: { authConfig: CloudViewerAuthConfig }) {
+  const { resolvedTheme } = useTheme(CLOUD_VIEWER_THEME_STORAGE_KEY);
+  const [bootstrap, setBootstrap] = useState<CloudNotebookListBootstrap | null>(() =>
+    loadCloudNotebookListBootstrap(),
+  );
+  const appSessionStatus = useCloudAppSessionStatus(bootstrap?.session ?? null);
+  const { authState, authRenewal, refreshAuthState } = useCloudPrototypeAuth(authConfig, {
+    appSessionRefreshFallback: true,
+    appSessionLoading: appSessionStatus.status === "loading",
+    appSession: appSessionStatus.session,
+  });
+  useCloudAppSessionBridge(
+    authState,
+    appSessionStatus.session,
+    appSessionStatus.status === "loading",
+    appSessionStatus.refreshAppSessionStatus,
+  );
+  const [listState, setListState] = useState<CloudNotebookListState>(() =>
+    initialCloudNotebookListState(authState, bootstrap),
+  );
+  const [refreshIndex, setRefreshIndex] = useState(0);
+  const [createState, setCreateState] = useState<"idle" | "starting">("idle");
+  const [createError, setCreateError] = useState<string | null>(null);
+  const [createFormOpen, setCreateFormOpen] = useState(false);
+  const [createTitle, setCreateTitle] = useState(() => defaultCloudNotebookTitle());
+  const [renameState, setRenameState] = useState<CloudNotebookRenameState | null>(null);
+  const [renameSavingId, setRenameSavingId] = useState<string | null>(null);
+  const [renameError, setRenameError] = useState<string | null>(null);
+  const hasExplicitAuth = authState.mode === "dev" || authState.mode === "oidc";
+  const hasAppSession = Boolean(appSessionStatus.session);
+  const signedIn = hasExplicitAuth || hasAppSession;
+  const dashboardModel = useMemo(
+    () => (listState.kind === "ready" ? projectCloudNotebookDashboard(listState.notebooks) : null),
+    [listState],
+  );
+
+  useEffect(() => {
+    applyDocumentTheme(resolvedTheme);
+  }, [resolvedTheme]);
+
+  useEffect(() => {
+    if (!signedIn) {
+      clearCachedCloudNotebookListFromWindow();
+      setListState({ kind: "signed_out" });
+      return;
+    }
+
+    const controller = new AbortController();
+    const cachedNotebooks =
+      readCachedCloudNotebookListFromWindow(authState) ?? bootstrap?.notebooks ?? null;
+    setListState(
+      cachedNotebooks ? { kind: "ready", notebooks: cachedNotebooks } : { kind: "loading" },
+    );
+    void (async () => {
+      try {
+        const response = await fetchCloudNotebookList(authState, controller.signal);
+        if (controller.signal.aborted) return;
+        if (!response.ok) {
+          throw await cloudResponseError(response, "Unable to list notebooks");
+        }
+        const body = (await response.json()) as unknown;
+        if (!isCloudNotebookListResponse(body)) {
+          throw new Error("Unable to list notebooks: response shape was invalid");
+        }
+        writeCachedCloudNotebookListToWindow(authState, body.notebooks);
+        setListState({ kind: "ready", notebooks: body.notebooks });
+      } catch (error) {
+        if (controller.signal.aborted) return;
+        setListState({
+          kind: "error",
+          message: error instanceof Error ? error.message : String(error),
+        });
+      }
+    })();
+
+    return () => {
+      controller.abort();
+    };
+  }, [authState, bootstrap, hasAppSession, refreshIndex, signedIn]);
+
+  const refreshList = () => {
+    setRefreshIndex((value) => value + 1);
+  };
+
+  const openCreateForm = () => {
+    if (!signedIn) {
+      return;
+    }
+    setCreateError(null);
+    setCreateTitle(defaultCloudNotebookTitle());
+    setCreateFormOpen(true);
+  };
+
+  const closeCreateForm = () => {
+    if (createState === "starting") {
+      return;
+    }
+    setCreateError(null);
+    setCreateFormOpen(false);
+  };
+
+  const createNotebook = async (event: FormEvent<HTMLFormElement>) => {
+    event.preventDefault();
+    if (!signedIn || createState === "starting") {
+      return;
+    }
+    const title = createTitle.trim() || defaultCloudNotebookTitle();
+    try {
+      setCreateError(null);
+      setCreateState("starting");
+      const response = await fetchWithCloudPrototypeAuth(
+        cloudNotebookCollectionEndpoint(),
+        {
+          method: "POST",
+          headers: {
+            Accept: "application/json",
+            "Content-Type": "application/json",
+          },
+          body: JSON.stringify({ title }),
+        },
+        cloudAuthWithScope(authState, "owner"),
+      );
+      if (!response.ok) {
+        throw await cloudResponseError(response, "Unable to create notebook");
+      }
+      const body = (await response.json()) as CloudNotebookCreateResponse;
+      if (body.ok !== true || typeof body.viewer_url !== "string") {
+        throw new Error("Unable to create notebook: response shape was invalid");
+      }
+      window.location.assign(body.viewer_url);
+    } catch (error) {
+      setCreateState("idle");
+      setCreateError(error instanceof Error ? error.message : String(error));
+    }
+  };
+
+  const openRenameForm = useCallback((notebook: CloudNotebookListItem) => {
+    setRenameError(null);
+    setRenameState({
+      notebookId: notebook.notebook_id,
+      title: notebook.title?.trim() ?? "",
+    });
+  }, []);
+
+  const closeRenameForm = useCallback(() => {
+    if (renameSavingId) {
+      return;
+    }
+    setRenameError(null);
+    setRenameState(null);
+  }, [renameSavingId]);
+
+  const saveNotebookTitle = async (event: FormEvent<HTMLFormElement>) => {
+    event.preventDefault();
+    if (!signedIn || !renameState || renameSavingId) {
+      return;
+    }
+
+    const notebookId = renameState.notebookId;
+    const nextTitle = renameState.title.trim();
+    try {
+      setRenameError(null);
+      setRenameSavingId(notebookId);
+      const response = await fetchWithCloudPrototypeAuth(
+        cloudNotebookCatalogEndpoint(notebookId),
+        {
+          method: "PATCH",
+          headers: {
+            Accept: "application/json",
+            "Content-Type": "application/json",
+          },
+          body: JSON.stringify({ title: nextTitle || null }),
+        },
+        cloudAuthWithScope(authState, "editor"),
+      );
+      if (!response.ok) {
+        throw await cloudResponseError(response, "Unable to rename notebook");
+      }
+      const body = (await response.json()) as CloudNotebookUpdateResponse;
+      if (body.ok !== true || body.notebook_id !== notebookId) {
+        throw new Error("Unable to rename notebook: response shape was invalid");
+      }
+      setListState((current) => {
+        if (current.kind !== "ready") {
+          return current;
+        }
+        const notebooks = current.notebooks.map((notebook) =>
+          notebook.notebook_id === notebookId
+            ? {
+                ...notebook,
+                title: body.title ?? null,
+                updated_at: body.updated_at ?? notebook.updated_at,
+                viewer_url: body.viewer_url ?? notebook.viewer_url,
+              }
+            : notebook,
+        );
+        writeCachedCloudNotebookListToWindow(authState, notebooks);
+        return {
+          kind: "ready",
+          notebooks,
+        };
+      });
+      setRenameState(null);
+    } catch (error) {
+      setRenameError(error instanceof Error ? error.message : String(error));
+    } finally {
+      setRenameSavingId(null);
+    }
+  };
+
+  const signOut = () => {
+    setBootstrap(null);
+    appSessionStatus.clearAppSessionStatus();
+    clearCachedCloudNotebookListFromWindow();
+    void clearCloudAppSession()
+      .catch((error: unknown) => {
+        console.warn("[notebook-cloud] app session clear failed", error);
+      })
+      .finally(appSessionStatus.refreshAppSessionStatus);
+    clearCloudPrototypeDevAuth(window.localStorage);
+    refreshAuthState();
+  };
+
+  const headerDetail =
+    authState.mode === "oidc" || authState.mode === "dev"
+      ? "Signed in"
+      : hasAppSession
+        ? "Signed in"
+        : authState.mode === "oidc_expired"
+          ? "Session expired"
+          : "Signed out";
+
+  return (
+    <main className="cloud-notebook-list-page">
+      <header className="cloud-notebook-list-header">
+        <div>
+          <a className="cloud-notebook-list-brand" href="/">
+            nteract
+          </a>
+          <h1>Notebook home</h1>
+          <p>{headerDetail}</p>
+        </div>
+        <div className="cloud-notebook-list-actions">
+          <button
+            type="button"
+            disabled={!signedIn || listState.kind === "loading"}
+            onClick={refreshList}
+          >
+            <RotateCcw aria-hidden="true" />
+            Refresh
+          </button>
+          <button
+            type="button"
+            disabled={!signedIn || createState === "starting"}
+            onClick={openCreateForm}
+          >
+            {createState === "starting" ? (
+              <Loader2 className="cloud-home-status-spinner" aria-hidden="true" />
+            ) : (
+              <FilePlus2 aria-hidden="true" />
+            )}
+            {createState === "starting" ? "Creating" : "New notebook"}
+          </button>
+          {hasExplicitAuth || hasAppSession ? null : (
+            <CloudNotebookSignInButton authConfig={authConfig} authState={authState} />
+          )}
+          {signedIn ? (
+            <button type="button" onClick={signOut}>
+              <LogOut aria-hidden="true" />
+              Sign out
+            </button>
+          ) : null}
+        </div>
+      </header>
+
+      {authRenewal.kind !== "idle" && !hasAppSession ? (
+        <div
+          className="cloud-notebook-list-banner"
+          data-kind={authRenewal.kind === "failed" ? "error" : "info"}
+          role={authRenewal.kind === "failed" ? "alert" : "status"}
+        >
+          {authRenewal.message}
+        </div>
+      ) : null}
+      {createError ? (
+        <div className="cloud-notebook-list-banner" data-kind="error" role="alert">
+          {createError}
+        </div>
+      ) : null}
+      {renameError ? (
+        <div className="cloud-notebook-list-banner" data-kind="error" role="alert">
+          {renameError}
+        </div>
+      ) : null}
+      {createFormOpen ? (
+        <form className="cloud-new-notebook-form" onSubmit={createNotebook}>
+          <label htmlFor="cloud-new-notebook-title">Notebook title</label>
+          <input
+            id="cloud-new-notebook-title"
+            type="text"
+            value={createTitle}
+            maxLength={160}
+            disabled={createState === "starting"}
+            onChange={(event) => setCreateTitle(event.currentTarget.value)}
+          />
+          <button type="submit" disabled={createState === "starting"}>
+            {createState === "starting" ? (
+              <Loader2 className="cloud-home-status-spinner" aria-hidden="true" />
+            ) : (
+              <FilePlus2 aria-hidden="true" />
+            )}
+            Create
+          </button>
+          <button type="button" disabled={createState === "starting"} onClick={closeCreateForm}>
+            Cancel
+          </button>
+        </form>
+      ) : null}
+
+      <section className="cloud-notebook-list-content" aria-label="Notebook list">
+        {listState.kind === "loading" ? (
+          <div className="cloud-notebook-list-state" data-kind="loading" role="status">
+            <Loader2 className="cloud-home-status-spinner" aria-hidden="true" />
+            <span>Loading notebooks</span>
+          </div>
+        ) : listState.kind === "signed_out" ? (
+          <div className="cloud-notebook-list-state" data-kind="signed-out">
+            <KeyRound aria-hidden="true" />
+            <span>Sign in to view notebooks.</span>
+          </div>
+        ) : listState.kind === "error" ? (
+          <div className="cloud-notebook-list-state" data-kind="error" role="alert">
+            <AlertCircle aria-hidden="true" />
+            <span>{listState.message}</span>
+          </div>
+        ) : listState.notebooks.length === 0 ? (
+          <div className="cloud-notebook-list-state" data-kind="empty">
+            <BookOpen aria-hidden="true" />
+            <span>No notebooks yet.</span>
+          </div>
+        ) : dashboardModel ? (
+          <CloudNotebookDashboard
+            model={dashboardModel}
+            canRename={signedIn}
+            renameState={renameState}
+            renameSavingId={renameSavingId}
+            onOpenRename={openRenameForm}
+            onCancelRename={closeRenameForm}
+            onRenameTitleChange={(title) =>
+              setRenameState((current) => (current ? { ...current, title } : current))
+            }
+            onSaveRename={saveNotebookTitle}
+          />
+        ) : (
+          <div className="cloud-notebook-list-state" data-kind="error" role="alert">
+            <AlertCircle aria-hidden="true" />
+            <span>Unable to project notebook dashboard.</span>
+          </div>
+        )}
+      </section>
+    </main>
+  );
+}
+
+function cloudNotebookListEndpoint(): string {
+  return new URL("api/n?limit=100", `${window.location.origin}/`).href;
+}
+
+function cloudNotebookCollectionEndpoint(): string {
+  return new URL("api/n", `${window.location.origin}/`).href;
+}
+
+function cloudNotebookCatalogEndpoint(notebookId: string): string {
+  return new URL(`api/n/${encodeURIComponent(notebookId)}`, `${window.location.origin}/`).href;
+}
+
+function cloudAuthWithScope(
+  authState: CloudPrototypeAuthState,
+  requestedScope: NonNullable<CloudPrototypeAuthState["requestedScope"]>,
+): CloudPrototypeAuthState {
+  return authState.requestedScope === requestedScope
+    ? authState
+    : {
+        ...authState,
+        requestedScope,
+      };
+}
+
+function initialCloudNotebookListState(
+  authState: CloudPrototypeAuthState,
+  bootstrap: CloudNotebookListBootstrap | null,
+): CloudNotebookListState {
+  const cachedNotebooks = readCachedCloudNotebookListFromWindow(authState) ?? bootstrap?.notebooks;
+  return cachedNotebooks ? { kind: "ready", notebooks: cachedNotebooks } : { kind: "loading" };
+}
+
+function fetchCloudNotebookList(
+  authState: CloudPrototypeAuthState,
+  signal: AbortSignal,
+): Promise<Response> {
+  if (authState.mode === "dev" || authState.mode === "oidc") {
+    return fetchWithCloudPrototypeAuth(
+      cloudNotebookListEndpoint(),
+      { headers: { Accept: "application/json" }, signal },
+      authState,
+    );
+  }
+  return fetch(cloudNotebookListEndpoint(), {
+    credentials: "same-origin",
+    headers: { Accept: "application/json" },
+    signal,
+  });
+}
+
+function readCachedCloudNotebookListFromWindow(
+  authState: CloudPrototypeAuthState,
+): CloudNotebookListItem[] | null {
+  const storage = cloudNotebookListCacheStorage();
+  return storage ? readCachedCloudNotebookList(storage, authState) : null;
+}
+
+function writeCachedCloudNotebookListToWindow(
+  authState: CloudPrototypeAuthState,
+  notebooks: CloudNotebookListItem[],
+): void {
+  const storage = cloudNotebookListCacheStorage();
+  if (!storage) {
+    return;
+  }
+  writeCachedCloudNotebookList(storage, authState, notebooks);
+}
+
+function clearCachedCloudNotebookListFromWindow(): void {
+  const storage = cloudNotebookListCacheStorage();
+  if (!storage) {
+    return;
+  }
+  clearCachedCloudNotebookList(storage);
+}
+
+function cloudNotebookListCacheStorage(): Storage | null {
+  try {
+    return window.sessionStorage;
+  } catch {
+    return null;
+  }
+}
+
+function defaultCloudNotebookTitle(now = new Date()): string {
+  const date = new Intl.DateTimeFormat(undefined, {
+    month: "short",
+    day: "numeric",
+    year: "numeric",
+  }).format(now);
+  const time = new Intl.DateTimeFormat(undefined, {
+    hour: "numeric",
+    minute: "2-digit",
+  }).format(now);
+  return `Notebook ${date} ${time}`;
+}
+
+function isCloudNotebookListResponse(value: unknown): value is CloudNotebookListResponse {
+  if (!value || typeof value !== "object") {
+    return false;
+  }
+  const candidate = value as Record<string, unknown>;
+  return candidate.ok === true && Array.isArray(candidate.notebooks);
+}
